Time: 1 ms (55.07%) | Memory: 16.4 MB (69.57%) - LeetSync

+++ 396-rotate-function/rotate-function.c
@@ -1,12 +1,12 @@
 int maxRotateFunction(int* nums, int numsSize) {
-    long f = 0, sum = 0;
+    int f = 0, sum = 0;
     for(int i = 0; i < numsSize; i++) {
         sum += nums[i];
-        f += (long) i * nums[i];
+        f += i * nums[i];
     }
-    long max = f;
+    int max = f;
     for(int k = 1; k < numsSize; k++) {
-        f = f + sum - ((long) numsSize * nums[numsSize - k]);
+        f = f + sum - (numsSize * nums[numsSize - k]);
         max = fmax(max, f);
     }
     return max;
